requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
paho-mqtt
opencv-python
ultralytics
numpy
